chore: remove .openapi-generator directory (#20)
The `.openapi-generator` directory is already mentioned in `.gitignore`,
but it had not been removed from the repository. In addition to removing
that directory, this updates the Makefiles so that the
`.openapi-generator` directory is explicitly removed whenever code is
regenerated.

We have to explicitly remove this directory because the
`semantic-release/git` plugin [intentionally does not respect
`.gitignore`](https://github.com/semantic-release/git/issues/345), so
these files will be re-committed on every release if they are not
explicitly removed during code generation.