feat: add support for Equinix Metal (#5)
This adds support for Equinix Metal by replicating the existing spec
patching and code generation processes from
[`metal-python`](https://github.com/equinix-labs/metal-python). This
ensures a smooth upgrade path for adopting `equinix-sdk-python` and
gaining access to other services supported by this SDK, and unblocks
https://github.com/packethost/packet-python/issues/136.

The code added here does not align with Equinix or OpenAPI standards and
is subject to change. In particular, handling of additional properties
does not follow OpenAPI rules and must be fixed in a future release.