chore: move Python version specification to a file (#30)
Previously we hard-coded the `python-version` in each CI workflow, which
has already resulted in version drift. This moves the Python version
specification to `.python-version` so that we can ensure that the same
Python version is used across all CI jobs. The presence of the
`.python-version` file also helps developers know what is the
recommended Python version to use for local SDK development.