chore: take 2, avoid committing .openapi-generator dir during release (#33)
The `.openapi-generator` directory came back in the most recent release,
despite the changes in #20. I _think_ that's because I didn't define the
globs in `assets` as a nested dir, so they get evaluated separately
instead of as a group. This pushes the asset globs down a level in an
attempt to get the `semantic-release/git` plugin to exclude the files we
want to exclude.

This change is based on the following text from [the plugin
docs](https://github.com/semantic-release/git?tab=readme-ov-file#assets):

> Each entry in the `assets` `Array` is globbed individually. A
[glob](https://github.com/micromatch/micromatch#matching-features) can
be a `String` (`"dist/**/*.js"` or `"dist/mylib.js"`) or an `Array` of
`String`s that will be globbed together (`["dist/**", "!**/*.css"]`).