chore(ci): only publish to PyPI if a new release was created (#38)
The semantic release GitHub Action will skip creating a new GitHub
release if there are no new releasable changes. This is expected, but
the release workflow assumed that there was always a new release to
publish to PyPI.

This PR updates the release workflow to move the PyPI publish steps to a
separate job that will only run if the release job successfully creates
a new release.