chore: adopt openapi-generator-ignore file to control output (#70)
This replaces the `modelTests` and `apiTests` settings in the generator
config file with a custom openapi-generator-ignore file so that we can
more explicitly and thoroughly control what files are produced by the
generator in each service package. Switching to an ignore file
eliminates the need for the `remove-unused` cleanup task, since we can
now avoid generating unwanted files in the first place.

As a side effect of this change, the auto-generated `metalv1` test files
have been removed. These test files were accidentally reintroduced when
the `generate` task in `Makefile.metalv1` was updated to allow the
`metalv1` code to continue to not follow OpenAPI standards.